api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.4",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "qstash>=3.2.0",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.38.0",
    "google-genai>=1.61.0",
    "stagehand>=3.5.0",
]
[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ============================================================================
# RUFF - Modern, fast linter and formatter (replaces black, isort, flake8)
# ============================================================================
[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

# Directories to exclude from linting/formatting
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable these rule categories
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modernize Python code)
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "B",      # flake8-bugbear (find likely bugs)
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "TRY",    # tryceratops (exception handling)
    "RUF",    # ruff-specific rules
]

# Disable specific rules that are too strict or conflict with FastAPI patterns
ignore = [
    "E501",    # Line too long (handled by formatter)
    "S101",    # Use of assert (fine in tests)
    "PLR0913", # Too many arguments (common in FastAPI endpoints)
    "TRY003",  # Avoid long messages outside exception class (too strict)
    "EM101",   # Exception must not use string literal (too strict)
    "EM102",   # Exception must not use f-string literal (too strict)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Ignore certain rules in test files
"tests/**/*.py" = [
    "S101",    # Use of assert
    "ARG",     # Unused function arguments (fixtures)
    "PLR2004", # Magic value comparisons
]
# Ignore import rules in __init__.py files
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

# ============================================================================
# MYPY - Static type checking
# ============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ============================================================================
# PYTEST - Testing configuration
# ============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
